Add basic flow collection FV in Goldmane
